Request permission for bluetooth on android 12
Using EasyPermissions to request BLUETOOTH_SCAN and BLUETOOTH_CONNECT on android 12. Otherwise, continue requesting location permissions.

Code is incomplete but functional if you follow the happy path.

Need to add checks for permissions before performing bluetooth actions. If the user denies the permissions to the app and then clicks a button, a crash or other undesirable behavior will happen.